+++ src/main/java/kitchenpos/orders/domain/model/OrderLineItem.java
@@ -1,15 +1,7 @@
 package kitchenpos.orders.domain.model;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.ForeignKey;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import jakarta.persistence.Transient;
+import jakarta.persistence.*;
+
 import kitchenpos.menus.domain.model.Menu;
 
 import java.math.BigDecimal;
