+++ src/main/java/kitchenpos/orders/domain/model/OrderTable.java
@@ -1,9 +1,6 @@
 package kitchenpos.orders.domain.model;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 
 import java.util.UUID;
